+++ src/RecoveryServices/RecoveryServices.Backup.Providers/Providers/IaasVmPsBackupProvider.cs
@@ -518,8 +518,8 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.ProviderModel
 
             Dictionary<UriEnums, string> uriDict = HelperUtils.ParseUri(rp.Id);
             string containerUri = HelperUtils.GetContainerUri(uriDict, rp.Id);
-            string securedVMOsDiskEncryptionSetId = ProviderData.ContainsKey(RestoreVMBackupItemParams.SecuredVMOsDiskEncryptionSetId) ?
-                ProviderData[RestoreVMBackupItemParams.SecuredVMOsDiskEncryptionSetId].ToString() : null;
+            string cVMOsDiskEncryptionSetId = ProviderData.ContainsKey(RestoreVMBackupItemParams.CVMOsDiskEncryptionSetId) ?
+                ProviderData[RestoreVMBackupItemParams.CVMOsDiskEncryptionSetId].ToString() : null;
 
             if (targetSubscriptionId == null || targetSubscriptionId == "") targetSubscriptionId = ServiceClientAdapter.SubscriptionId;
 
@@ -649,9 +649,9 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.ProviderModel
                 }                
             }
 
-            if (!string.IsNullOrWhiteSpace(securedVMOsDiskEncryptionSetId))
+            if (!string.IsNullOrWhiteSpace(cVMOsDiskEncryptionSetId))
             {
-                restoreRequest.SecuredVMDetails = new SecuredVMDetails(securedVMOsDiskEncryptionSetId);
+                restoreRequest.SecuredVMDetails = new SecuredVMDetails(cVMOsDiskEncryptionSetId);
             }
 
             if (restoreType == "OriginalLocation") // replace existing