+++ src/RecoveryServices/RecoveryServices/ChangeLog.md
@@ -18,7 +18,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
-* Added new optional parameter SecuredVMOsDiskEncryptionSetId in Restore-AzRecoveryServicesBackupItem.
+* Added new optional parameter CVMOsDiskEncryptionSetId in Restore-AzRecoveryServicesBackupItem.
 
 ## Version 7.7.2
 * Added fix for resume protection with AzureFileshare.